refactor(main/graphql): add float types - cart - price - review

+++ src/main/graphql/types/cart.ts
@@ -6,8 +6,8 @@ extend type Query {
 
 type Cart {
   id: UUID!
-  total: Int!
-  subTotal: Int!
+  total: Float!
+  subTotal: Float!
   createdAt: DateTime!
   updatedAt: DateTime
