+++ src/main/graphql/types/review.ts
@@ -2,7 +2,7 @@ export const reviewTypeDefs = `#graphql
 type Review {
   id: UUID!
   ticketId: UUID!
-  value: Int!
+  value: Float!
   createdAt: DateTime!
 }
 `
